Renamed the example package to examples.
