+++ indexer-examples/src/main/java/org/apache/maven/indexer/examples/BasicUsageExample.java
@@ -1,4 +1,4 @@
-package org.apache.maven.indexer.example;
+package org.apache.maven.indexer.examples;
 
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
